Fix reconnection: persist device_address, clear connectedPeripheral on iOS fail, extract hook
- Android: save BLE MAC to GlassesStore on GATT connect so direct-GATT
  reconnection actually works (was dead code — device_address never saved)
- Android: persist device_address in handleDeviceReady and clear on forget
- iOS: clear connectedPeripheral in didFailToConnect so reconnect guard
  doesn't stall after a failed connection attempt
- iOS: fix GlassesStore key "deviceAddress" → "device_address" for
  consistency with Android/RN, uncomment device_address save, clear on forget
- iOS: persist peripheral UUID as device_address on BLE connect
- RN: extract duplicated wasSearching debounce + nativeLinkBusy into
  useSearchingState hook shared by CompactDeviceStatus and DeviceStatus
