+++ mobile/modules/core/android/src/main/java/com/mentra/core/CoreManager.kt
@@ -1036,6 +1036,7 @@ class CoreManager {
         // save the default_wearable now that we're connected:
         Bridge.saveSetting("default_wearable", defaultWearable)
         Bridge.saveSetting("device_name", deviceName)
+        Bridge.saveSetting("device_address", deviceAddress)
     }
 
     private fun handleG1Ready() {
@@ -1380,8 +1381,10 @@ class CoreManager {
         // Clear state
         defaultWearable = ""
         deviceName = ""
+        deviceAddress = ""
         Bridge.saveSetting("default_wearable", "")
         Bridge.saveSetting("device_name", "")
+        Bridge.saveSetting("device_address", "")
     }
 
     fun findCompatibleDevices(deviceModel: String) {
